Update README repository url and setup instructions
The git repository URL in README has been updated to the correct one as previous URL was not accessible. Also, an additional step has been added for setup procedure to configure the OPENAI_API_KEY. This is essential for the application to function properly and hence was included in the instructions.