AAdd model picker, streaming, and redesign the UI as well as dependency updates.
This commit introduces several new UI components including badge, form, select and tooltip. Additionally, it extends the ModalType in use-modal-store to include 'aimodel'. The global styling rules have also been updated to enhance layout and visual presentation. Also, modifications are made in Layout.tsx and package-lock.json for better functionality and proper dependency resolution. This also introduces streaming, allowing for longer responses, as well as choosing the model for the AI to use and even a full redesign!